done refactor evaluations table to ag-grid

+++ frontend-angular/src/app/components/pages/evaluation/evaluation.component.ts
@@ -2,7 +2,7 @@ import { Component, ViewChild } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTable, MatTableDataSource } from '@angular/material/table';
-import { IEvaluationEntity } from '../../../../../../interfaces';
+import { IEvaluationEntity, IGoalEntity } from '../../../../../../interfaces';
 import { Subscription } from 'rxjs';
 import { EvaluationService } from 'src/app/services/evaluation.service';
 import { UtilityService } from 'src/app/services/utility.service';
@@ -10,6 +10,8 @@ import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute } from '@angular/router';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { AddEditEvaluationComponent } from '../../dialogs/add-edit/add-edit-evaluation/add-edit-evaluation.component';
+import { ColDef, GridOptions, ISetFilterParams, NewValueParams } from 'ag-grid-community';
+import { AgGridService, MyMenuItem } from 'src/app/services/ag-grid.service';
 
 @Component({
   selector: 'app-evaluation',
@@ -17,43 +19,103 @@ import { AddEditEvaluationComponent } from '../../dialogs/add-edit/add-edit-eval
   styleUrls: ['./evaluation.component.scss']
 })
 export class EvaluationComponent {
-  public canAdd!: boolean;
-  public canEditDelete!: boolean;
-  @ViewChild(MatPaginator) paginator!: MatPaginator;
-  @ViewChild(MatSort) sort!: MatSort;
-  @ViewChild(MatTable) table!: MatTable<IEvaluationEntity>;
-  public dataSource!: MatTableDataSource<IEvaluationEntity>;
-  public columnsKeys!: string[];
+  public canAdd: boolean = this.ut.userHasAny('Admin', 'Teacher');
+  public canEditDelete: boolean = this.ut.userHasAny('Admin', 'Teacher', 'HeadOfDepartment');
+  public selectedItem?: IEvaluationEntity;
+  public quickFilter: string = '';
+  public isPrinting: boolean = false;
+  /**don't use `rowData` 'cause child has goals for `rowData`*/
+  public goalItsEvaluations: IGoalEntity | undefined;
   private sub: Subscription = new Subscription();
 
-  constructor(public service: EvaluationService, public ut: UtilityService, private dialog: MatDialog, private route: ActivatedRoute) {
+  private onCellValueChange = async (e: NewValueParams<IEvaluationEntity>) => {
+    try {
+      await this.service.patch(e.data.id, { [e.colDef.field as keyof IEvaluationEntity]: e.newValue });
+      this.ut.showSnackbar('Edited successfully')
+    } catch (e) {
+      if (this.goalItsEvaluations)
+        await this.service.fetchGoalItsEvaluations(this.goalItsEvaluations.id).catch(() => { });
+      this.gridOptions?.api?.refreshCells();
+    }
+  }
+
+  /**
+* @see [ag-grid.service](./../../../services/ag-grid.service.ts) for more information of how to set the columnDef properties.
+*/
+  public columnDefs: (ColDef<IEvaluationEntity>)[] = [
+    {
+      field: 'description',
+      headerName: 'Description',
+      type: 'long',
+      onCellValueChanged: this.onCellValueChange,
+    },
+    {
+      field: 'rate',
+      headerName: 'Rate',
+      type: 'enum',
+      filterParams: {
+        values: ['continual', 'excellent'],
+        valueFormatter: v => v.value == 'continual' ? this.ut.translate('Continual') : this.ut.translate('Excellent'),
+      } as ISetFilterParams<IGoalEntity, string>,
+      valueGetter: (v) => v.data?.rate == 'excellent' ? this.ut.translate('Excellent') : this.ut.translate('Continual'),
+    },
+    {
+      field: 'mainstream',
+      headerName: 'Mainstream',
+      type: 'long',//filterParams in init
+      onCellValueChanged: this.onCellValueChange,
+    },
+    {
+      field: 'note',
+      headerName: 'Note',
+      type: 'long',
+      onCellValueChanged: this.onCellValueChange,
+    },
+    {
+      field: 'evaluationDatetime',
+      headerName: 'Assign Date',
+      type: 'fromNow',
+      onCellValueChanged: this.onCellValueChange,
+    },
+    {
+      field: 'teacher.person.name',
+      headerName: 'By Teacher',
+      type: 'long',
+    },
+  ];
+
+  private menuItems: MyMenuItem<IEvaluationEntity>[] = [
+    {
+      name: 'Delete',
+      icon: `<mat-icon _ngcontent-glk-c62="" color="warn" role="img" class="mat-icon notranslate mat-warn material-icons mat-ligature-font" aria-hidden="true" data-mat-icon-type="font">delete</mat-icon>`,
+      action: (v) => this.deleteDialog(v),
+      tooltip: 'Delete the selected evaluation',
+      disabled: !this.canEditDelete,
+    },
+  ];
+
+
+
+  constructor(public service: EvaluationService, public ut: UtilityService,
+    private dialog: MatDialog, private route: ActivatedRoute,
+    public agGrid: AgGridService) {
   }
 
 
 
   ngOnInit(): void {
-    this.canAdd = this.ut.userHasAny('Admin', 'Teacher');
-    this.canEditDelete = this.ut.userHasAny('Admin', 'Teacher', 'HeadOfDepartment');
-    this.columnsKeys = JSON.parse(sessionStorage.getItem('evaluations table') ?? 'null') ?? ['description', 'rate', 'mainstream', 'evaluationDatetime', 'note', 'teacher', 'menu'];
     this.route.paramMap.subscribe({
       next: async params => {
         let goalId = params.get('id');
         if (typeof goalId === 'string')
-          await this.service.fetchGoalItsEvaluations(+goalId, true).catch(() => { });
+          this.sub.add(this.service.goalItsEvaluations.subscribe(async v => {
+            if (v && v.id == +(goalId as string))
+              this.goalItsEvaluations = v;
+            else await this.service.fetchGoalItsEvaluations(+(goalId as string));
+          }));
         else this.ut.errorDefaultDialog(undefined, "Sorry, there was a problem fetching the goal's evaluations. Please try again later or check your connection.");
       },
     });
-
-    this.dataSource = new MatTableDataSource<IEvaluationEntity>();
-    this.sub.add(this.service.goalItsEvaluations.subscribe({
-      next: (v) => {
-        if (v == null)
-          return;
-        this.dataSource.data = v.evaluations;
-        if (this.table)
-          this.table.renderRows();
-      }
-    }));
 
     this.sub.add(this.ut.user.subscribe(v => {
       this.canAdd = this.ut.userHasAny('Admin', 'Teacher');
@@ -61,23 +123,20 @@ export class EvaluationComponent {
     }));
   }
 
-  ngAfterViewInit(): void {
-    this.dataSource.sort = this.sort;
-    this.dataSource.paginator = this.paginator;
+  applySearch(event: Event) {
+    this.quickFilter = (event.target as HTMLInputElement).value;
   }
 
-
-  applyFilter(event: Event) {
-    const filterValue = (event.target as HTMLInputElement).value;
-    this.dataSource.filter = filterValue.trim().toLowerCase();
-
-    if (this.dataSource.paginator)
-      this.dataSource.paginator.firstPage();
+  printTable() {
+    this.agGrid.printTable(this.gridOptions, v => this.isPrinting = v);
   }
 
-  drop(event: CdkDragDrop<string[]>) {
-    moveItemInArray(this.columnsKeys, event.previousIndex, event.currentIndex);
-    sessionStorage.setItem('evaluations table', JSON.stringify(this.columnsKeys));
+  /**Before adding any attribute. Check if it exist in commonGridOptions. So, no overwrite happen!  */
+  public gridOptions: GridOptions<IEvaluationEntity> = {
+    ...this.agGrid.commonGridOptions('evaluations table', this.columnDefs, this.canEditDelete,
+      this.menuItems, this.printTable, (item) => { this.addEdit(item) },
+    ),
+    onRowClicked: (v) => this.selectedItem = v.data,
   }
 
   /** @param data is either a evaluation to be Edit. Or goalId to be Add */
@@ -93,19 +152,22 @@ export class EvaluationComponent {
         });
   }
 
-  deleteDialog(evaluation: IEvaluationEntity) {
-    this.ut.showMsgDialog({
-      content: this.ut.translate('You are about to delete the evaluation with description: \"') + evaluation.description+ this.ut.translate("\" permanently."),
-      type: 'confirm',
-      buttons: [{ color: 'primary', type: 'Cancel' }, { color: 'warn', type: 'Delete' }]
-    }).afterClosed().subscribe(async (v) => {
-      if (v === 'Delete') {
-        try {
-          await this.service.delete(evaluation.id, true);
-          this.ut.showSnackbar('The evaluation has been deleted successfully.');
-        } catch (e) { }
-      }
-    })
+  deleteDialog(evaluation?: IEvaluationEntity) {
+    if (evaluation == null)
+      this.ut.showSnackbar(undefined);
+    else
+      this.ut.showMsgDialog({
+        content: this.ut.translate('You are about to delete the evaluation with description: \"') + evaluation.description + this.ut.translate("\" permanently."),
+        type: 'confirm',
+        buttons: [{ color: 'primary', type: 'Cancel' }, { color: 'warn', type: 'Delete' }]
+      }).afterClosed().subscribe(async (v) => {
+        if (v === 'Delete') {
+          try {
+            await this.service.delete(evaluation.id, true);
+            this.ut.showSnackbar('The evaluation has been deleted successfully.');
+          } catch (e) { }
+        }
+      })
   }
 
   ngOnDestroy(): void {
